+++ ABPlayer/Sources/Plugins/PluginProtocol.swift
@@ -1,0 +1,10 @@
+import SwiftUI
+
+@MainActor
+protocol Plugin: Identifiable {
+  var id: String { get }
+  var name: String { get }
+  var icon: String { get }
+  
+  func open()
+}